requirements.txt
Django==4.1.2
Markdown==3.1.1
psycopg2>=2.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-ckeditor==5.9.0
Pillow==9.2.0
django-environ
# JWT
pyjwt==1.7.1
django-rest-swagger==2.2.0
django-cors-headers